remove JavaClassBuilderService
