+++ src/main/kotlin/org/jetbrains/research/testspark/display/TestCasePanelFactory.kt
@@ -28,9 +28,9 @@ import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.data.UIContext
 import org.jetbrains.research.testspark.data.llm.JsonEncoding
 import org.jetbrains.research.testspark.display.custom.IJProgressIndicator
+import org.jetbrains.research.testspark.helpers.JavaClassBuilderHelper
 import org.jetbrains.research.testspark.helpers.LLMHelper
 import org.jetbrains.research.testspark.services.ErrorService
-import org.jetbrains.research.testspark.services.JavaClassBuilderService
 import org.jetbrains.research.testspark.services.LLMSettingsService
 import org.jetbrains.research.testspark.services.ReportLockingService
 import org.jetbrains.research.testspark.services.TestCaseDisplayService
@@ -455,9 +455,7 @@ class TestCasePanelFactory(
         WriteCommandAction.runWriteCommandAction(project) {
             project.service<ErrorService>().clear()
             val code = testSuitePresenter.toString(testSuite)
-            testCase.testName =
-                project.service<JavaClassBuilderService>()
-                    .getTestMethodNameFromClassWithTestCase(testCase.testName, code)
+            testCase.testName = JavaClassBuilderHelper.getTestMethodNameFromClassWithTestCase(testCase.testName, code)
             testCase.testCode = code
 
             // update numbers
@@ -517,7 +515,7 @@ class TestCasePanelFactory(
 
         val newTestCase = TestProcessor(project)
             .processNewTestCase(
-                "${project.service<JavaClassBuilderService>().getClassFromTestCaseCode(testCase.testCode)}.java",
+                "${JavaClassBuilderHelper.getClassFromTestCaseCode(testCase.testCode)}.java",
                 testCase.id,
                 testCase.testName,
                 testCase.testCode,
@@ -662,8 +660,7 @@ class TestCasePanelFactory(
      */
     private fun updateTestCaseInformation() {
         testCase.testName =
-            project.service<JavaClassBuilderService>()
-                .getTestMethodNameFromClassWithTestCase(testCase.testName, languageTextField.document.text)
+            JavaClassBuilderHelper.getTestMethodNameFromClassWithTestCase(testCase.testName, languageTextField.document.text)
         testCase.testCode = languageTextField.document.text
     }
 }
